@@ -9,7 +9,7 @@
     "clean": "rimraf ./dist; mkdir -p dist/public",
     "dev": "concurrently \"npm run dev:app\" \"npm run dev:server\"",
     "dev:admin": "concurrently \"npm run dev:admin-app\" \"npm run dev:server\"",
-    "dev:app": "npm run prebuild:app && concurrently \"ng build --watch --app time-client-universal\" \"ng build --watch\" ",
+    "dev:app": "npm run prebuild:app && ng build --watch",
     "dev:server": "npm run build:server && concurrently \"npm run watch:server\" \"nodemon dist/server.js\"",
     "dev:admin-app": "npm run prebuild:admin-app && ng build --watch --app time-admin",
     "prebuild:app": "node ./build-utils/time-client.prebuild",
@@ -19,8 +19,8 @@
     "build:prod": "npm run clean; npm i; npm run build:prod-app && npm run build:server",
     "build:staging": "npm run clean; npm i; npm run build:staging-app && npm run build:server",
     "build:admin": "npm run build:admin-app",
-    "build:app": "npm run prebuild:app && ng build; ng build --app time-client-universal",
-    "build:prod-app": "npm run prebuild:app && ng build --prod --no-aot --environment=prod && ng build --app time-client-universal --prod --no-aot --environment=prod",
+    "build:app": "npm run prebuild:app && ng build",
+    "build:prod-app": "npm run prebuild:app && ng build --prod --no-aot --environment=prod",
     "build:staging-app": "npm i; npm run prebuild:app && ng build --prod --no-aot --environment=staging && ng build --app time-client-universal --prod --no-aot --environment=staging",
     "build:admin-app": "npm run prebuild:admin-app && ng build --app time-admin",
     "build:server": "npm run prebuild:api && webpack --config time-api/config/webpack.server-config.js",
@@ -34,7 +34,7 @@
     "e2e": "protractor ./time-client/protractor.conf.js"
   },
   "dependencies": {
-    "@nguniversal/express-engine": "^1.0.0-beta.2",
+    "@nguniversal/express-engine": "^5.0.0-beta.5",
     "JSONStream": "1.3.1",
     "aws-sdk": "^2.153.0",
     "bcrypt-nodejs": "0.0.3",
@@ -80,23 +80,23 @@
     "simple-encryptor": "^1.1.0",
     "stripe": "^5.3.0",
     "web-animations-js": "^2.3.1",
-    "webpack": "3.4.1",
+    "xmlhttprequest": "^1.8.0",
     "zone.js": "0.8.14"
   },
   "devDependencies": {
-    "@angular/animations": "4.4.3",
-    "@angular/cli": "^1.5.0",
-    "@angular/common": "4.4.3",
-    "@angular/compiler": "4.4.3",
-    "@angular/compiler-cli": "4.4.3",
-    "@angular/core": "4.4.3",
-    "@angular/forms": "4.4.3",
-    "@angular/http": "4.4.3",
-    "@angular/language-service": "4.4.3",
-    "@angular/platform-browser": "4.4.3",
-    "@angular/platform-browser-dynamic": "4.4.3",
-    "@angular/platform-server": "4.4.3",
-    "@angular/router": "4.4.3",
+    "@angular/animations": "5.0.3",
+    "@angular/cli": "1.5.4",
+    "@angular/common": "5.0.3",
+    "@angular/compiler": "5.0.3",
+    "@angular/compiler-cli": "5.0.3",
+    "@angular/core": "5.0.3",
+    "@angular/forms": "5.0.3",
+    "@angular/http": "5.0.3",
+    "@angular/language-service": "5.0.3",
+    "@angular/platform-browser": "5.0.3",
+    "@angular/platform-browser-dynamic": "5.0.3",
+    "@angular/platform-server": "5.0.3",
+    "@angular/router": "5.0.3",
     "@types/bluebird": "3.5.8",
     "@types/body-parser": "1.16.4",
     "@types/express": "4.0.36",
@@ -134,7 +134,6 @@
     "karma-jasmine-html-reporter": "0.2.2",
     "less-loader": "4.0.2",
     "ng-click-outside": "2.4.1",
-    "ngx-toastr": "^6.1.4",
     "ngx-uploader": "3.3.4",
     "postcss-loader": "1.3.3",
     "postcss-url": "5.1.2",
@@ -147,13 +146,14 @@
     "source-map-loader": "0.2.0",
     "style-loader": "0.13.1",
     "stylus-loader": "3.0.1",
-    "to-string-loader": "^1.1.5",
+    "to-string-loader": "1.1.5",
     "ts-loader": "2.3.1",
     "ts-node": "3.2.0",
     "tslint": "5.3.2",
     "typegoose": "^3.7.0",
-    "typescript": "^2.3.4",
+    "typescript": "2.4.2",
     "url-loader": "0.5.7",
+    "webpack": "3.4.1",
     "webpack-dev-server": "2.5.1"
   }
 }
