@@ -3,20 +3,21 @@ import { BrowserModule } from '@angular/platform-browser'
 import { RouterModule } from '@angular/router'
 
 import { AppComponent } from './app.component'
+import { SharedModule } from './shared/shared.module'
 import { ShopModule } from './shop/shop.module'
 import { SiteModule } from './site/site.module'
 
 @NgModule({
-    declarations: [
-        AppComponent,
-    ],
     imports: [
-        // BrowserModule,
         // ANGULAR UNIVERSAL
-        BrowserModule.withServerTransition({appId: 'time-client-universal'}),
+        BrowserModule.withServerTransition({ appId: 'time-client-universal' }),
         RouterModule.forRoot([]),
+        SharedModule.forRoot(),
         SiteModule,
         ShopModule,
+    ],
+    declarations: [
+        AppComponent,
     ],
     providers: [],
     bootstrap: [AppComponent],
