@@ -6,6 +6,7 @@ import {
     Router,
     RouterStateSnapshot,
 } from '@angular/router'
+import { Observable } from 'rxjs/Observable'
 
 import { AppRoutes } from '../../../../constants/app-routes'
 import { RouteStateService } from './route-state.service'
@@ -20,18 +21,21 @@ export class AuthGuardService implements CanActivate, CanActivateChild {
         private routeState: RouteStateService,
     ) {}
 
-    public canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean {
-        if (this.userService.user) {
-            return true
-        }
-        else {
-            const redirectTo = this.routeState.previousUrl || AppRoutes.Login
-            this.router.navigate([redirectTo])
-            return false
-        }
+    public canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> {
+        return this.userService.user$
+            .map(user => {
+                if (user) {
+                    return true
+                }
+                else {
+                    const redirectTo = this.routeState.previousUrl || AppRoutes.Login
+                    this.router.navigate([redirectTo])
+                    return false
+                }
+            })
     }
 
-    public canActivateChild(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean {
+    public canActivateChild(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> {
         return this.canActivate(route, state)
     }
 
