@@ -1,6 +1,7 @@
 import { Inject, Injectable } from '@angular/core'
-import { DOCUMENT, Title } from '@angular/platform-browser'
-import { AppConfig } from '@time/app-config'
+import { DOCUMENT } from '@angular/platform-browser'
+
+import { UtilService } from './util.service'
 
 @Injectable()
 export class SeoService {
@@ -10,20 +11,26 @@ export class SeoService {
     private robots: HTMLElement
 
     constructor(
-        private titleService: Title,
+        private util: UtilService,
         @Inject(DOCUMENT) private DOM: any,
     ) {
-        this.headElement = this.DOM.querySelector('head')
+        if (!this.util.isServerApp()) {
+            this.headElement = this.DOM.querySelector('head')
+        }
+        else {
+            this.headElement = new HTMLElement()
+        }
+
         this.metaDescription = this.getOrCreateMetaElement('description')
         this.robots = this.getOrCreateMetaElement('robots')
     }
 
     public getTitle(): string {
-        return this.titleService.getTitle()
+        return this.util.getTitle()
     }
 
     public setTitle(newTitle: string): void {
-        this.titleService.setTitle(newTitle + ' | ' + AppConfig.brand_name)
+        this.util.setTitle(newTitle)
     }
 
     public getMetaDescription(): string {
@@ -48,14 +55,19 @@ export class SeoService {
      * @returns {HTMLElement}
      */
     private getOrCreateMetaElement(name: string): HTMLElement {
-        let el: HTMLElement
-        el = this.DOM.querySelector('meta[name=' + name + ']')
-        if (el == null) {
-            el = this.DOM.createElement('meta')
-            el.setAttribute('name', name)
-            this.headElement.appendChild(el)
+        if (!this.util.isServerApp()) {
+            let el: HTMLElement
+            el = this.DOM.querySelector('meta[name=' + name + ']')
+            if (el == null) {
+                el = this.DOM.createElement('meta')
+                el.setAttribute('name', name)
+                this.headElement.appendChild(el)
+            }
+            return el
         }
-        return el
+        else {
+            return new HTMLElement()
+        }
     }
 
 }
