@@ -1,27 +1,16 @@
 import { ModuleWithProviders, NgModule } from '@angular/core'
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations'
 
-/*
- * EXTERNAL MODULES
- */
-import { ToastrModule } from 'ngx-toastr'
-
-/*
- * FEATURE MODULES
- */
+// FEATURE MODULES
 import { TimeFormsModule } from '@time/common/ng-modules/forms'
 import { TimeHttpModule } from '@time/common/ng-modules/http'
 import { TimeUiModule } from '@time/common/ng-modules/ui'
 
-/*
- * COMPONENTS
- */
+// COMPONENTS
 import { LoginComponent } from './components/login/login.component'
 import { SiteLocatorComponent } from './components/site-locator/site-locator.component'
 
-/*
- * SERVICES
- */
+// SERVICES
 import {
     AuthGuardService,
     CartService,
@@ -37,12 +26,6 @@ import {
 @NgModule({
     imports: [
         BrowserAnimationsModule,
-        ToastrModule.forRoot({
-            timeOut: 6000,
-            positionClass: 'toast-top-center',
-            maxOpened: 1,
-            preventDuplicates: true,
-        }),
         TimeHttpModule.forRoot(),
         TimeUiModule,
         TimeFormsModule,
@@ -59,7 +42,6 @@ import {
         LoginComponent,
     ],
 })
-
 export class SharedModule {
     public static forRoot(): ModuleWithProviders {
         return {
