@@ -3,7 +3,6 @@ import { plugin, prop, Ref } from 'typegoose'
 
 import { Attribute } from './attribute'
 import { BaseApiModel } from './base-api-model'
-import { MixedTypeValue } from './mixed-type-value'
 
 @plugin(findOrCreate)
 export class AttributeValue extends BaseApiModel<AttributeValue> {
@@ -11,19 +10,7 @@ export class AttributeValue extends BaseApiModel<AttributeValue> {
     @prop() public name: string
     @prop() public slug: string
     @prop() public description: string
-    @prop() public value: MixedTypeValue
-
-    public getValue(): any {
-        const value = this.value
-        let theValue: any
-
-        Object.keys(value).forEach((key) => {
-            if (typeof value[key] !== 'undefined') {
-                theValue = value[key]
-            }
-        })
-        return theValue
-    }
+    @prop() public value: any
 }
 
 export const AttributeValueModel = new AttributeValue().getModelForClass(AttributeValue)
