@@ -1,9 +1,20 @@
 import { prop } from 'typegoose'
 
+import { CurrencyEnum } from '@time/common/constants/currency'
+
 export class EasypostRate {
+    public get getDate(): {
+        created_at: Date
+        updated_at: Date
+    } {
+        return {
+            created_at: new Date(this.created_at),
+            updated_at: new Date(this.updated_at),
+        }
+    }
     @prop() public readonly created_at: string
     @prop() public readonly updated_at: string
-    @prop() public readonly mode: string /* "test" or "production" */
+    @prop({ enum: ['test', 'production'] }) public readonly mode: string /* "test" or "production" */
     @prop() public readonly service: string /* service level/name */
     @prop() public readonly carrier: string /* name of carrier */
     @prop() public readonly carrier_account_id: string /* ID of the CarrierAccount record used to generate this rate */
@@ -13,7 +24,7 @@ export class EasypostRate {
     @prop() public readonly retail_rate: string /* the retail rate is the in-store rate given with no account */
     @prop() public readonly retail_currency: string /* currency for the retail rate */
     @prop() public readonly list_rate: string /* the list rate is the non-negotiated rate given for having an account with the carrier */
-    @prop() public readonly list_currency: string /* currency for the list rate */
+    @prop({ enum: CurrencyEnum }) public readonly list_currency: string /* currency for the list rate */
     @prop() public readonly delivery_days: number /* delivery days for this service */
     @prop() public readonly delivery_date: string /* date for delivery */
     @prop() public readonly delivery_date_guaranteed: boolean /* indicates if delivery window is guaranteed (true) or not (false) */
