@@ -1,92 +1,117 @@
 import {
-    ChangeDetectionStrategy,
-    ChangeDetectorRef,
     Component,
-    EventEmitter,
     Input,
+    OnDestroy,
+    OnInit,
     Output,
 } from '@angular/core'
+import { Observable } from 'rxjs/Observable'
+import { Subscription } from 'rxjs/Subscription'
 
+import { AppConfig } from '@time/app-config'
 import { IModalData } from '../../../../models/interfaces/ui/modal-data'
+import { platform } from '../../utils/platform'
+import { timeout } from '../../utils/timeout'
 
 @Component({
     selector: 'time-modal',
     template: `
-        <div *ngIf="data" class="modal-container" [hidden]="!isShowing" [ngClass]="{'showing': isShowing}">
-            <div class="modal-darken" [ngStyle]="{'opacity': fadeIn ? 1 : 0}" (click)="show = false"></div>
-            <div *ngIf="data" class="modal-inner-wrapper">
-                <div class="modal-inner" [ngStyle]="{'opacity': fadeIn ? 1 : 0}">
+        <div *ngIf="isShowing && data"
+             class="modal-container">
+            <div class="modal-darken"
+                 [ngStyle]="{'opacity': isFadedIn ? 1 : 0}"
+                 (click)="cancel()"></div>
+            <div class="modal-inner-wrapper">
+                <div class="modal-inner"
+                     [ngStyle]="{'opacity': isFadedIn ? 1 : 0}">
                     <div class="modal-title-bar">
                         <h2>{{ data.title }}</h2>
-                        <button class="blank-btn modal-close-btn" (click)="show = false"><img alt="close modal" src="static/images/x-dark.svg"></button>
+                        <button (click)="cancel()" class="blank-btn modal-close-btn">
+                            <img alt="close modal" [src]="_config.client_url + '/images/x-dark.svg'">
+                        </button>
                     </div>
                     <div class="modal-content" *ngIf="data.body">
                         <ng-container *compile="data.body; context: data.context"></ng-container>
                     </div>
                     <div class="modal-actions">
-                        <button (click)="cancel()" class="btn-cancel modal-cancel">{{ data.cancelText || defaultCancelText }}</button>
-                        <button *ngIf="data.okText" (click)="data.okClick()" class="btn-action modal-ok">{{ data.okText }}</button>
+                        <button (click)="cancel()" class="btn-cancel modal-cancel">
+                            {{ data.cancelText || defaultCancelText }}
+                        </button>
+                        <button *ngIf="data.okText" (click)="data.okClick()" class="btn-action modal-ok">
+                            {{ data.okText }}
+                        </button>
                     </div>
                 </div>
             </div>
         </div>
     `,
-    changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class TimeModalComponent {
-    @Input() public data: IModalData
-    @Input() public closeCallback: () => void
-    @Output() public showChange = new EventEmitter<boolean>()
+export class TimeModalComponent implements OnInit, OnDestroy {
+    @Input() public data$: Observable<IModalData>
+    @Input() public closeCallback?: () => void
+    public data: IModalData = null
     public defaultCancelText = "Cancel"
-    public isShowing: boolean
-    private isTransitioning: boolean
-    public fadeIn: boolean
+    public isShowing = false
+    public isFadedIn = false
+    public _config = AppConfig
+    private isTransitioning = false
     private modalInner: HTMLElement
 
-    @Input() public get show(): boolean {
-        return this.isShowing
+    public subscriptions = {
+        data: undefined,
+        initFadeInTimer: undefined,
+        endFadeInTimer: undefined,
+        endFadeOutTimer: undefined,
     }
-    public set show(isShowing: boolean) {
-        if (this.isTransitioning) return
-        if (!this.data) return
-        this.isTransitioning = true
 
-        if (isShowing) {
+    public ngOnInit() {
+        this.subscriptions.data = this.data$.subscribe((data) => this.show(data))
+    }
+
+    private show(data: IModalData) {
+        if (this.isTransitioning) return
+        this.isTransitioning = true
+        this.data = data
+
+        if (data) {
             this.isShowing = true
-            setTimeout(() => {
-                this.updateYPos(window.scrollY)
-                this.fadeIn = true
-            }, 10)
-            setTimeout(() => {
+            this.subscriptions.initFadeInTimer = timeout(10).subscribe(() => {
+                if (platform.isBrowser()) {
+                    this.updateYPos(window.scrollY)
+                }
+                this.isFadedIn = true
+            })
+            this.subscriptions.endFadeInTimer = timeout(400).subscribe(() => {
                 this.isTransitioning = false
-            }, 400)
-            this.showChange.emit(this.isShowing)
+            })
         }
         else {
-            this.fadeIn = false
-            setTimeout(() => {
+            this.isFadedIn = false
+            this.subscriptions.endFadeOutTimer = timeout(400).subscribe(() => {
                 this.isShowing = false
-                this.showChange.emit(this.isShowing)
                 this.isTransitioning = false
                 if (this.closeCallback) {
                     this.closeCallback()
                 }
-            }, 400)
+            })
         }
     }
 
-    constructor(
-        private cd: ChangeDetectorRef,
-    ) { }
+    public ngOnDestroy() {
+        Object.keys(this.subscriptions).forEach(key => {
+            if (this.subscriptions[key] as Subscription) {
+                this.subscriptions[key].unsubscribe()
+            }
+        })
+    }
 
     public updateYPos(scrollTop: number): any {
         this.modalInner = <HTMLElement>document.querySelector(".showing .modal-inner-wrapper")
         if (!this.modalInner) return
         this.modalInner.style.top = (scrollTop / 10 + 11).toString() + 'rem'
-        this.cd.detectChanges()
     }
 
     public cancel() {
-        this.show = false
+        this.show(null)
     }
 }
